Move test to operations package
Also now using constants instead of magic strings
